requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
Deprecated==1.2.12
distro==1.5.0
dnspython==1.16.0
et-xmlfile==1.1.0
idna==2.10
jsonpickle==2.0.0
multidict==5.1.0
openpyxl==3.0.7
pymongo==3.11.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
yookassa==2.1.2
aioschedule==0.5.2
loguru
pypng